Hotfix 01 on index.html
Ok, so I noticed a small mistake in the code, involving the fifth and sixth dimensions having some of the same variables. I fixed that in this little hotfix. I am planning on adding a new number display system to add to this game and other projects by me. It involves exponents. Lots and lots of them. This may break the limit of 2^1024, but I'm not telling you how. See ya!
- The Incremental Nerd