Removed the "Thanos Snap" bug
Don't blame me for this, but some Discord user (The one who possibly thought of this game) discovered a bug in which the lower half of the dimensions' display wasn't working properly. It turns out that some idiot didn't follow the pattern of the dimension stats, which starts at 0 for the first dimension, and 9 for the 10th dimension. That idiot made the stats ID for dimensions 1 higher than it normally was intended to be. I called that the Thanos Snap bug, and I reverse-snapped it out of existence. Peace out!
-The Incremental Nerd