Added our contributions
So today, I decided to create a .md file to let the player know about who helped with the game and what we did, because how are we supposed to remember what we did with the game? We can't just forget about our contributions to this game! We gotta jot it down somewhere, right??? Well, I'll see ya real soon. 
-The Incremental Nerd